[Issue #96] Add test cart counter. Add all steps to one file.
